python/packages/knowledge/pyproject.toml
[project]
name = "knowledge"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Alex Whiteside", email = "alex.whiteside1@gmail.com" }
]
dependencies = [
    "lancedb>=0.11.0",
    "pandas>=2.2.2",
    "python-arango>=8.0.0",
    "langchain>=0.2.11",
    "ollama>=0.3.1",
    "langchain-ollama>=0.1.0",
    "langchain-community>=0.2.10",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/knowledge"]
